+++ storybook/cypress/e2e/dropdown-menu.cy.ts
@@ -88,9 +88,8 @@ describe("Dropdown menu - anchors", () => {
     cy.get("@menuitems").should("be.visible");
   });
 
-  // Todo: #3034
-  it.skip("shift-tab should cycle through options and button in reverse order", { browser: "!firefox" }, () => {
-    cy.get("@button").focus().click();
+  it("shift-tab should cycle through options and button in reverse order", { browser: "!firefox" }, () => {
+    cy.get("@button").should("be.visible").focus().click();
 
     cy.realPress("Tab");