+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Task
 * Packages: Dependencies updates ([#2977](https://github.com/dso-toolkit/dso-toolkit/issues/2977))
+* Dropdown Menu: flaky test `shift-tab should cycle through options and button in reverse order` ([#3034](https://github.com/dso-toolkit/dso-toolkit/issues/3034))
 
 ## 🪼 Release 69.2.1 - 2025-03-21